High level architecture, Vault backend start and CLI start.
The command line interface is decoupled from the certmaestro library itself.

The config is read from a .ini file, it can be reloaded and saved.

The Backend config only knows about their own settings.

The Client (only CLI yet) needs to know the necessary configuration to initialize the backend.

The backend can add configuration parameters to it's own section,
but can't save or delete or reload it.

section_param is a descriptor which can define default values. Every configuration parameter
can be find in the backend's Config class.
## To make a commit, type your commit message and press SUPER-ENTER. To cancel
## the commit, close the window. To sign off on the commit, press SUPER-S.
##
## You may also reference or close a GitHub issue with this commit.  To do so,
## type `#` followed by the `tab` key.  You will be shown a list of issues
## related to the current repo.  You may also type `owner/repo#` plus the `tab`
## key to reference an issue in a different GitHub repo.