Remove validation from Param level
actually validation at Param level doesn't make sense. It's better to check everything in __init__,
so it's in one place and if anything is wrong, it will be clear when something tries to instantiate
the backend class. Later maybe it will make sense to separate validation, but it's a grey zone
because is it part of the validation that we can't connect to the Vault server for example? The user
probably only care if his/her settings are all good.

The exception from this is validate_setup because that's sort of dry run? We don't want to mess up
an already valid and working configuration so we need to check first without side effect to the
production system.